Changes to be committed: 	new file:   __pycache__/level_1.cpython-314.pyc 	new file:   __pycache__/level_2.cpython-314.pyc 	new file:   font_test.py 	modified:   level_1.py 	modified:   level_2.py 	new file:   main_screen.py
Used GEMINI CLI to fix pygame font loading issue. It was a system issue,
but prompt sculpting got GEMINI to fix the problem. I added a
main_screen.py file to easily swap between levels, and this will make it
so much easier to add new levels in the future. All code in level_1 and
level_2 is now under a function that will be called when its button is
pressed in the main menu_screen.

UP AND COMING:
Button in all levels that will bring back to the main menu_screen.

Better color pallete that is easy on the eyes

FAR IN THE FUTURE:
PNGS instead of pygame rects