Changes to be committed: 	new file:   level_3.py 	new file:   universal.py
I added a new universal.py file just so that I can have a copy and a
paste for all of my files. Technically it is not completely universal
because menu screens wont need player initiation, but you get the point.
Later I might make a "menu copy paste".py file, but probably not because
it is easy to initiate a file on my own.